Fixed int8_t to int16_t

+++ milestone05/src/arch/x86_64/keyboard.c
@@ -248,7 +248,7 @@ void modify_state(uint8_t code){
 }
 
 void keyboard_read(int irq, int err, void *data){
-   int8_t code;
+   int16_t code;
    code = ps2_read(PS2_DATA);
 
    if (code < 0)
